Extract helper method to helper module
